+++ templates/data/options.py
@@ -50,6 +50,13 @@ exts = [
 # if False: use QRZ's default name format
 qrz_only_nickname = True
 
+# enable a command that provides a link to add the bot to a server
+enable_invite_cmd = True
+
+# the default permissions for the bot, to be included in the invite link for ?invite
+# this probably does not need to be changed
+invite_perms = 67488832
+
 # Either "time", "random", or "fixed" (first item in statuses)
 status_mode = "fixed"